# Optimize Speedy machine.ctrl