--- a/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SExpr.scala
+++ b/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SExpr.scala
@@ -9,8 +9,11 @@ package com.daml.lf.speedy
   * This reduces the number of binding forms by moving update and scenario
   * expressions into builtins.
   */
+import java.util
+
 import com.daml.lf.language.Ast._
 import com.daml.lf.data.Ref._
+import com.daml.lf.value.{Value => V}
 import com.daml.lf.speedy.SValue._
 import com.daml.lf.speedy.Speedy._
 import com.daml.lf.speedy.SError._
@@ -24,8 +27,7 @@ import java.util.ArrayList
   * - all update and scenario operations converted to builtin functions.
   */
 sealed abstract class SExpr extends Product with Serializable {
-  def execute(machine: Machine): Ctrl
-
+  def execute(machine: Machine): Unit
   @SuppressWarnings(Array("org.wartremover.warts.Any"))
   override def toString: String =
     productPrefix + productIterator.map(SExpr.prettyPrint).mkString("(", ",", ")")
@@ -39,8 +41,8 @@ object SExpr {
     * https://en.wikipedia.org/wiki/De_Bruijn_index
     */
   final case class SEVar(index: Int) extends SExpr {
-    def execute(machine: Machine): Ctrl = {
-      CtrlValue(machine.getEnv(index))
+    def execute(machine: Machine): Unit = {
+      machine.returnValue(machine.getEnv(index))
     }
   }
 
@@ -51,27 +53,29 @@ object SExpr {
       ref: SDefinitionRef,
       var cached: Option[(SValue, List[Location])],
   ) extends SExpr {
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       machine.lookupVal(this)
     }
   }
 
   /** Reference to a builtin function */
   final case class SEBuiltin(b: SBuiltin) extends SExpr {
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       /* special case for nullary record constructors */
       b match {
         case SBRecCon(id, fields) if b.arity == 0 =>
-          CtrlValue(SRecord(id, fields, new ArrayList()))
+          machine.returnValue(SRecord(id, fields, new ArrayList()))
         case _ =>
-          Ctrl.fromPrim(PBuiltin(b), b.arity)
+          machine.returnValue(SPAP(PBuiltin(b), new util.ArrayList[SValue](), b.arity))
       }
     }
   }
 
   /** A pre-computed value, usually primitive literal, e.g. integer, text, boolean etc. */
   final case class SEValue(v: SValue) extends SExpr {
-    def execute(machine: Machine): Ctrl = CtrlValue(v)
+    def execute(machine: Machine): Unit = {
+      machine.returnValue(v)
+    }
   }
 
   object SEValue extends SValueContainer[SEValue]
@@ -80,9 +84,9 @@ object SExpr {
     * evaluates to a builtin or a closure.
     */
   final case class SEApp(fun: SExpr, args: Array[SExpr]) extends SExpr with SomeArrayEquals {
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       machine.pushKont(KArg(args))
-      CtrlExpr(fun)
+      machine.ctrl = fun
     }
   }
 
@@ -91,7 +95,7 @@ object SExpr {
     * can be written against this simplified expression type.
     */
   final case class SEAbs(arity: Int, body: SExpr) extends SExpr {
-    def execute(machine: Machine): Ctrl =
+    def execute(machine: Machine): Unit =
       crash("unexpected SEAbs, expected SEMakeClo")
   }
 
@@ -110,7 +114,7 @@ object SExpr {
       extends SExpr
       with SomeArrayEquals {
 
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       def convertToSValues(fv: Array[Int], getEnv: Int => SValue) = {
         val sValues = new Array[SValue](fv.length)
         var i = 0
@@ -122,15 +126,15 @@ object SExpr {
       }
 
       val sValues = convertToSValues(fv, machine.getEnv)
-      Ctrl.fromPrim(PClosure(body, sValues), arity)
+      machine.returnValue(SPAP(PClosure(body, sValues), new util.ArrayList[SValue](), arity))
     }
   }
 
   /** Pattern match. */
   final case class SECase(scrut: SExpr, alts: Array[SCaseAlt]) extends SExpr with SomeArrayEquals {
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       machine.pushKont(KMatch(alts))
-      CtrlExpr(scrut)
+      machine.ctrl = scrut
     }
 
     override def toString: String = s"SECase($scrut, ${alts.mkString("[", ",", "]")})"
@@ -155,7 +159,7 @@ object SExpr {
     * with later expressions possibly referring to earlier.
     */
   final case class SELet(bounds: Array[SExpr], body: SExpr) extends SExpr with SomeArrayEquals {
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       // Pop the block once we're done evaluating the body
       machine.pushKont(KPop(bounds.size))
 
@@ -167,7 +171,7 @@ object SExpr {
         val b = bounds(bounds.size - i)
         machine.pushKont(KPushTo(machine.env, b))
       }
-      CtrlExpr(bounds.head)
+      machine.ctrl = bounds.head
     }
   }
 
@@ -193,9 +197,9 @@ object SExpr {
     * variable of the machine. When commit is begun the location is stored in 'commitLocation'.
     */
   final case class SELocation(loc: Location, expr: SExpr) extends SExpr {
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       machine.pushLocation(loc)
-      CtrlExpr(expr)
+      machine.ctrl = expr
     }
   }
 
@@ -207,9 +211,29 @@ object SExpr {
     * to access the value returned from 'body'.
     */
   final case class SECatch(body: SExpr, handler: SExpr, fin: SExpr) extends SExpr {
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       machine.pushKont(KCatch(handler, fin, machine.env.size))
-      CtrlExpr(body)
+      machine.ctrl = body
+    }
+  }
+
+  /** When we fetch a contract id from upstream we cannot crash in the upstream
+    * calls. Rather, we set the control to this expression and then crash when executing.
+    */
+  final case class SEWronglyTypeContractId(
+      acoid: V.AbsoluteContractId,
+      expected: TypeConName,
+      actual: TypeConName,
+  ) extends SExpr {
+    def execute(machine: Machine): Unit = {
+      throw DamlEWronglyTypedContract(acoid, expected, actual)
+    }
+  }
+
+  final case class SEImportValue(value: V[V.ContractId]) extends SExpr {
+    def execute(machine: Machine): Unit = {
+      machine.importValue(value)
+//>>>>>>> ddfe844ec... Optimize Speedy machine.ctrl
     }
   }
 
@@ -263,7 +287,7 @@ object SExpr {
 
     import SEBuiltinRecursiveDefinition._
 
-    def execute(machine: Machine): Ctrl = {
+    def execute(machine: Machine): Unit = {
       val body = ref match {
         case Reference.FoldL => foldLBody
         case Reference.FoldR => foldRBody