--- a/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SExpr.scala
+++ b/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/SExpr.scala
@@ -42,7 +42,7 @@ object SExpr {
     */
   final case class SEVar(index: Int) extends SExpr {
     def execute(machine: Machine): Unit = {
-      machine.returnValue(machine.getEnv(index))
+      machine.returnValue = machine.getEnv(index)
     }
   }
 
@@ -62,11 +62,11 @@ object SExpr {
   final case class SEBuiltin(b: SBuiltin) extends SExpr {
     def execute(machine: Machine): Unit = {
       /* special case for nullary record constructors */
-      b match {
+      machine.returnValue = b match {
         case SBRecCon(id, fields) if b.arity == 0 =>
-          machine.returnValue(SRecord(id, fields, new ArrayList()))
+          SRecord(id, fields, new ArrayList())
         case _ =>
-          machine.returnValue(SPAP(PBuiltin(b), new util.ArrayList[SValue](), b.arity))
+          SPAP(PBuiltin(b), new util.ArrayList[SValue](), b.arity)
       }
     }
   }
@@ -74,7 +74,7 @@ object SExpr {
   /** A pre-computed value, usually primitive literal, e.g. integer, text, boolean etc. */
   final case class SEValue(v: SValue) extends SExpr {
     def execute(machine: Machine): Unit = {
-      machine.returnValue(v)
+      machine.returnValue = v
     }
   }
 
@@ -126,7 +126,7 @@ object SExpr {
       }
 
       val sValues = convertToSValues(fv, machine.getEnv)
-      machine.returnValue(SPAP(PClosure(body, sValues), new util.ArrayList[SValue](), arity))
+      machine.returnValue = SPAP(PClosure(body, sValues), new util.ArrayList[SValue](), arity)
     }
   }