--- a/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Speedy.scala
+++ b/daml-lf/interpreter/src/main/scala/com/digitalasset/daml/lf/speedy/Speedy.scala
@@ -21,16 +21,16 @@ import scala.util.control.NoStackTrace
 
 object Speedy {
 
-  // Avoid chains of calls to returnDepth, which will blow the JVM stack
-  val limitReturnDepth = 1000
-
   /** The speedy CEK machine. */
   final case class Machine(
-      /* So we can limit chains of calls to returnDepth. */
-      var returnDepth: Int,
-      //var maxReturnDepth: Int, //Instrumentation
-      /* The control is what the machine should be evaluating. */
+      /* The control is what the machine should be evaluating. If this is not
+       * null, then `returnValue` must be null.
+       */
       var ctrl: SExpr,
+      /* `returnValue` contains the result once the expression in `ctrl` has
+       * been fully evaluated. If this is not null, then `ctrl` must be null.
+       */
+      var returnValue: SValue,
       /* The environment: an array of values */
       var env: Env,
       /* Kont, or continuation specifies what should be done next
@@ -147,17 +147,20 @@ object Speedy {
         // note: exception handler is outside while loop
         try {
           while (!isFinal) {
-            //if (returnDepth > maxReturnDepth) maxReturnDepth = returnDepth;
-            returnDepth = 0
-            ctrl.execute(this) // make a single step
-          }
-          ctrl match {
-            case SEValue(value) => {
-              //println(s"maxReturnDepth = $maxReturnDepth")
-              result = SResultFinalValue(value) //stop
+            if (returnValue != null) {
+              val value = returnValue
+              returnValue = null
+              popKont.execute(value, this)
+            } else {
+              val expr = ctrl
+              ctrl = null
+              expr.execute(this)
             }
-            case _ =>
-              throw SErrorCrash(s"Unexpected ctrl on final machine $ctrl")
+          }
+          if (returnValue != null) {
+            result = SResultFinalValue(returnValue)
+          } else {
+            throw SErrorCrash(s"Unexpected ctrl on final machine $ctrl")
           }
         } catch {
           case SpeedyHungry(res: SResult) => result = res //stop
@@ -194,7 +197,7 @@ object Speedy {
       eval.cached match {
         case Some((v, stack_trace)) =>
           pushStackTrace(stack_trace)
-          returnValue(v)
+          returnValue = v
 
         case None =>
           val ref = eval.ref
@@ -224,68 +227,10 @@ object Speedy {
 
     /** Returns true when the machine has finished evaluation.
       * The machine is considered final when the kont stack
-      * is empty, and the control expression is a value
+      * is empty, and we have a `returnValue`, and hence `ctrl` is null.
       */
     def isFinal(): Boolean = {
-      if (kontStack.isEmpty) {
-        ctrl match {
-          case SEValue(_) => true
-          case _ => false
-        }
-      } else
-        false
-    }
-
-    // Return a value to the waiting continutaion
-    def returnValue(value: SValue): Unit = {
-      returnDepth += 1
-      if (kontStack.isEmpty) {
-        ctrl = SEValue(value) // The final resulting value
-      } else {
-        if (returnDepth < limitReturnDepth) {
-          val kont = popKont()
-          kont.execute(value, this)
-        } else {
-          ctrl = SEValue(value) // force the current step to end
-        }
-      }
-    }
-
-    /**
-      Unwind all KPop and KLocation continuations from the kont-stack.
-      For KPop - we do the required popEnv.
-      Execute the first continutaion found which is not KPop/KLocation.
-      Or, if kont-stack becomes empty, set the machine.ctrl to the final value.
-
-      This function has the same semantics as `returnValue` but avoids consuming JVM stack
-      for sequences of KPop/KLocation.  We only use this function from the `execute`
-      method of KPop/KLocation, so not every call to returnDepth has the pay the price of
-      the extra matching on the Kont class which is done here.
-      Even so, this function has a small negative impact on performance.
-      */
-    def unwindKontStackThenReturnValue(value: SValue): Unit = {
-      var kont: Kont = null
-      var again: Boolean = true
-      while (again) {
-        if (kontStack.isEmpty) {
-          again = false
-        } else {
-          popKont() match {
-            case KLocation(_) => ()
-            case KPop(count) =>
-              popEnv(count)
-            case properContinution => {
-              again = false
-              kont = properContinution
-            }
-          }
-        }
-      }
-      if (kont != null) {
-        kont.execute(value, this)
-      } else {
-        ctrl = SEValue(value) // The final resulting value
-      }
+      kontStack.isEmpty && returnValue != null
     }
 
     def enterFullyAppliedFunction(prim: Prim, args: util.ArrayList[SValue]): Unit = {
@@ -323,16 +268,20 @@ object Speedy {
       if (!isFinal) {
         crash("toSValue: machine not final")
       } else {
-        ctrl match {
-          case SEValue(v) => v
-          case _ => crash("machine did not evaluate to a value")
-        }
+        returnValue
       }
 
     def print(count: Int) = {
       println(s"Step: $count")
-      println("Control:")
-      println(s"  ${ctrl}")
+      if (returnValue != null) {
+        println("Control: null")
+        println("Return:")
+        println(s"  ${returnValue}")
+      } else {
+        println("Control:")
+        println(s"  ${ctrl}")
+        println("Return: null")
+      }
       println("Environment:")
       env.forEach { v =>
         println("  " + v.toString)
@@ -432,7 +381,7 @@ object Speedy {
                     id.packageId,
                     pkg => {
                       compiledPackages = pkg
-                      returnValue(go(value))
+                      returnValue = go(value)
                     }
                   ))
             }
@@ -451,12 +400,12 @@ object Speedy {
                     id.packageId,
                     pkg => {
                       compiledPackages = pkg
-                      returnValue(go(value))
+                      returnValue = go(value)
                     }
                   ))
             }
         }
-      returnValue(go(value))
+      returnValue = go(value)
     }
 
   }
@@ -472,9 +421,8 @@ object Speedy {
         globalCids: Set[V.AbsoluteContractId]
     ) =
       Machine(
-        returnDepth = 0,
-        //maxReturnDepth = 0,
         ctrl = null,
+        returnValue = null,
         env = emptyEnv,
         kontStack = new util.ArrayList[Kont](128),
         lastLocation = None,
@@ -585,7 +533,7 @@ object Speedy {
   final case class KPop(count: Int) extends Kont {
     def execute(v: SValue, machine: Machine) = {
       machine.popEnv(count)
-      machine.unwindKontStackThenReturnValue(v)
+      machine.returnValue = v
     }
   }
 
@@ -636,7 +584,7 @@ object Speedy {
         machine.enterFullyAppliedFunction(prim, args)
       } else {
         // args.size < arity (we already dealt with args.size > args in Karg)
-        machine.returnValue(SPAP(prim, args, arity))
+        machine.returnValue = SPAP(prim, args, arity)
       }
     }
   }
@@ -719,9 +667,8 @@ object Speedy {
 
       machine.ctrl = altOpt
         .getOrElse(throw DamlEMatchError(s"No match for $v in ${alts.toList}"))
-        .body,
+        .body
     }
-
   }
 
   /** Push the evaluated value to the array 'to', and start evaluating the expression 'next'.
@@ -743,11 +690,11 @@ object Speedy {
     * accessed. In older compilers which did not use the builtin record and struct
     * updates this solves the blow-up which would happen when a large record is
     * updated multiple times. */
-  final case class KCacheVal(eval: SEVal, stack_trace: List[Location]) extends Kont {
-    def execute(v: SValue, machine: Machine) = {
+  final case class KCacheVal(v: SEVal, stack_trace: List[Location]) extends Kont {
+    def execute(sv: SValue, machine: Machine) = {
       machine.pushStackTrace(stack_trace)
-      eval.cached = Some((v, stack_trace))
-      machine.returnValue(v)
+      v.cached = Some((sv, stack_trace))
+      machine.returnValue = sv
     }
   }
 
@@ -765,7 +712,7 @@ object Speedy {
   /** A location frame stores a location annotation found in the AST. */
   final case class KLocation(location: Location) extends Kont {
     def execute(v: SValue, machine: Machine) = {
-      machine.unwindKontStackThenReturnValue(v)
+      machine.returnValue = v
     }
   }