--- a/daml-lf/scenario-interpreter/src/test/scala/com/digitalasset/daml/lf/speedy/ScenarioRunnerTest.scala
+++ b/daml-lf/scenario-interpreter/src/test/scala/com/digitalasset/daml/lf/speedy/ScenarioRunnerTest.scala
@@ -7,7 +7,6 @@ import com.daml.lf.PureCompiledPackages
 import com.daml.lf.data.{Ref, Time}
 import com.daml.lf.language.Ast
 import com.daml.lf.language.Ast.ScenarioGetParty
-import com.daml.lf.speedy.SExpr.SEValue
 import org.scalatest._
 import org.scalatest.concurrent.ScalaFutures
 
@@ -25,7 +24,8 @@ class ScenarioRunnerTest extends AsyncWordSpec with Matchers with ScalaFutures {
       )
       val sr = ScenarioRunner(m, _ + "-XXX")
       sr.run()
-      m.ctrl shouldBe SEValue(SValue.SParty(Ref.Party.assertFromString("foo-bar-XXX")))
+      m.ctrl shouldBe null
+      m.returnValue shouldBe SValue.SParty(Ref.Party.assertFromString("foo-bar-XXX"))
     }
   }