--- a/daml-script/runner/src/main/scala/com/digitalasset/daml/lf/engine/script/Runner.scala
+++ b/daml-script/runner/src/main/scala/com/digitalasset/daml/lf/engine/script/Runner.scala
@@ -309,7 +309,7 @@ class Runner(
       }
 
     def run(expr: SExpr): Future[SValue] = {
-      machine.ctrl = Speedy.CtrlExpr(expr)
+      machine.ctrl = expr
       stepToValue()
         .fold(Future.failed, Future.successful)
         .flatMap {