@hurryabit 2020-05-06
This was part of https://github.com/digital-asset/daml/issues/5764.